# Conflicts:
#	src/client/java/minicraft/core/Renderer.java
#	src/client/java/minicraft/entity/furniture/DungeonChest.java
#	src/client/java/minicraft/entity/mob/AirWizard.java
#	src/client/java/minicraft/entity/mob/Creeper.java
#	src/client/java/minicraft/entity/mob/Knight.java
#	src/client/java/minicraft/entity/mob/ObsidianKnight.java
#	src/client/java/minicraft/entity/mob/Skeleton.java
#	src/client/java/minicraft/entity/mob/Slime.java
#	src/client/java/minicraft/entity/mob/Snake.java
#	src/client/java/minicraft/entity/mob/Zombie.java
#	src/client/java/minicraft/entity/particle/SandParticle.java
#	src/client/java/minicraft/gfx/MinicraftImage.java
#	src/client/java/minicraft/gfx/Screen.java
#	src/client/java/minicraft/gfx/SpriteAnimation.java
#	src/client/java/minicraft/gfx/SpriteLinker.java
#	src/client/java/minicraft/item/BookItem.java
#	src/client/java/minicraft/item/TileItem.java
#	src/client/java/minicraft/item/TorchItem.java
#	src/client/java/minicraft/item/WateringCanItem.java
#	src/client/java/minicraft/level/tile/BossWallTile.java
#	src/client/java/minicraft/level/tile/farming/CarrotTile.java
#	src/client/java/minicraft/level/tile/farming/HeavenlyBerriesTile.java
#	src/client/java/minicraft/level/tile/farming/HellishBerriesTile.java
#	src/client/java/minicraft/level/tile/farming/PotatoTile.java
#	src/client/java/minicraft/level/tile/farming/TomatoTile.java
#	src/client/java/minicraft/level/tile/farming/WheatTile.java
#	src/client/java/minicraft/screen/ResourcePackDisplay.java
#	src/client/java/minicraft/screen/SkinDisplay.java

+++ src/client/java/minicraft/item/WateringCanItem.java
@@ -32,7 +32,7 @@ public class WateringCanItem extends Item {
 	private static final SpriteManager.SpriteLink particleSprite = new SpriteManager.SpriteLink.SpriteLinkBuilder(
 		SpriteManager.SpriteType.Entity, "glint").createSpriteLink();
 
-	private static final SpriteManager.SpriteLink[] spriteSplash = new SpriteManager.SpriteLink[]{
+	private static final SpriteManager.SpriteLink[] spriteSplash = new SpriteManager.SpriteLink[] {
 		new SpriteManager.SpriteLink.SpriteLinkBuilder(SpriteManager.SpriteType.Entity, "splash_0").createSpriteLink(),
 		new SpriteManager.SpriteLink.SpriteLinkBuilder(SpriteManager.SpriteType.Entity, "splash_1").createSpriteLink(),
 		new SpriteManager.SpriteLink.SpriteLinkBuilder(SpriteManager.SpriteType.Entity, "splash_2").createSpriteLink(),
@@ -88,7 +88,7 @@ public class WateringCanItem extends Item {
 						level.add(new Particle((int) x, (int) y, 120 + random.nextInt(21) - 40, particleSprite));
 					}
 					if (random.nextInt(60) == 0) { // Small chance for growing flowers
-						level.setTile(xt, yt, Tiles.get(2), random.nextInt(2));
+						level.setTile(xt, yt, Tiles.get((short) 2), random.nextInt(2));
 					}
 				}
